+++ src/pybamm/models/full_battery_models/lithium_ion/__init__.py
@@ -24,9 +24,9 @@ from .basic_dfn_composite import BasicDFNComposite
 from .Yang2017 import Yang2017
 from .mpm import MPM
 from .msmr import MSMR
-from .basic_ecm_split_OCV import ECMsplitOCV
+from .basic_splitOCVR import SplitOCVR
 
 __all__ = ['Yang2017', 'base_lithium_ion_model', 'basic_dfn',
            'basic_dfn_composite', 'basic_dfn_half_cell', 'basic_spm', 'dfn',
            'electrode_soh', 'electrode_soh_half_cell', 'mpm', 'msmr',
-           'newman_tobias', 'spm', 'spme', 'ecm_split_ocv']
+           'newman_tobias', 'spm', 'spme', 'basic_splitOCVR']