+++ tests/unit/test_models/test_full_battery_models/test_lithium_ion/test_splitOCVR.py
@@ -4,12 +4,12 @@
 import pybamm
 
 
-class TestECMSplitOCV:
+class TestSplitOCVR:
     def test_ecmsplitocv_well_posed(self):
-        model = pybamm.lithium_ion.ECMsplitOCV()
+        model = pybamm.lithium_ion.SplitOCVR()
         model.check_well_posedness()
 
     def test_get_default_quick_plot_variables(self):
-        model = pybamm.lithium_ion.ECMsplitOCV()
+        model = pybamm.lithium_ion.SplitOCVR()
         variables = model.default_quick_plot_variables
         assert "Current [A]" in variables